+++ packages/dds/tree/src/core/tree/index.ts
@@ -108,6 +108,9 @@ export {
 	makeDetachedNodeId,
 	offsetDetachId,
 	emptyDelta,
+	areDetachedNodeIdsEqual,
+	deltaFieldMapHasVisibleChanges,
+	deltaFieldChangesHaveVisibleChanges,
 } from "./deltaUtil.js";
 
 export {